cp_mes_django/requirements/development.txt
# Development dependencies
-r base.txt

# Debug toolbar
django-debug-toolbar==4.3.0

# Development server
watchdog==4.0.0

# Interactive shell
ipython==8.22.1
ipdb==0.13.13

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==24.0.0

# Code quality
flake8==7.0.0
black==24.2.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.6
djangorestframework-stubs==3.14.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
